Don't clear csv by default at the beginning of a workflow.

+++ silk-plugins/silk-plugins-csv/src/main/scala/org/silkframework/plugins/dataset/csv/CsvSink.scala
@@ -42,18 +42,20 @@ class CsvSink(val resource: WritableResource, settings: CsvSettings) extends Dat
     * Makes sure that the next write will start from an empty dataset.
     */
   override def clear()(implicit userContext: UserContext): Unit = {
-    val resourceFile = new File(resource.path).getAbsoluteFile
-    val resourcePath = resourceFile.toPath
-    val crcFile = new File(resourcePath.getParent.toFile, s".${resourcePath.getFileName.toString}.crc")
-    resource.delete()
-    // Delete CRC file if exists
-    try {
-      if (crcFile.exists() && crcFile.canWrite && crcFile.isFile) {
-        crcFile.delete()
+    if(settings.clearBeforeExecution) {
+      val resourceFile = new File(resource.path).getAbsoluteFile
+      val resourcePath = resourceFile.toPath
+      val crcFile = new File(resourcePath.getParent.toFile, s".${resourcePath.getFileName.toString}.crc")
+      resource.delete()
+      // Delete CRC file if exists
+      try {
+        if (crcFile.exists() && crcFile.canWrite && crcFile.isFile) {
+          crcFile.delete()
+        }
+      } catch {
+        case e: IOException =>
+          log.warning("IO exception occurred when deleting CRC file: " + e.getMessage)
       }
-    } catch {
-      case e: IOException =>
-        log.warning("IO exception occurred when deleting CRC file: " + e.getMessage)
     }
   }
 }
